% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine_train_vectors.R
\name{combine_train_vectors}
\alias{combine_train_vectors}
\title{Combine OASIS training vectors}
\usage{
combine_train_vectors(list_of_train_vectors)
}
\arguments{
\item{list_of_train_vectors}{a list of outputs from the oasis_trian_vectors}
}
\value{
This function reutrns a dataframe for use with the oasis_training function.
}
\description{
This function combines multiple oasis training dataframes produced by the oasis_trian_vectors function and produces a larger dataframe
that can be used in the oasis_training function.  The multiple dataframes should be passed into the function in list form.
}
\examples{
\dontrun{
study_1_dataframe <- oasis_train_vectors(flair = FLAIR_1 ,t1 = T1_1 , t2 = T2_1, pd = PD_1, gold_standard = GOLD_STANDARD_1)
study_2_dataframe <- oasis_train_vectors(flair = FLAIR_2 ,t1 = T1_2 , t2 = T2_2, pd = PD_2, gold_standard = GOLD_STANDARD_2)
multistudy_dataframe <- combine_train_vectors(list(study_1_dataframe, study_2_dataframe))
  }
}

