% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correct_image_dim.R
\name{correct_image_dim}
\alias{correct_image_dim}
\title{Image Dimmension Correction}
\usage{
correct_image_dim(image, dim = 3)
}
\arguments{
\item{image}{volume of class nifti}

\item{dim}{scalar value of desired image dimmension}
}
\value{
Reutrns a volume of class nifti of desired dimmension.
}
\description{
This function takes an image and drops dimmensions until the volume is a user specified dimmension.
}
\examples{
\dontrun{
flair <- readNIfTI('path/to/flair', reorient = FALSE)
correct_image_dim <- function(flair, dim = 3) }
}

