% Generated by roxygen2 (4.0.0): do not edit by hand
\name{courseEval}
\alias{courseEval}
\title{Function for taking in a raw course evaluation dataset from OASIS
and outputting a table of interim values that can be later formatted for use
with knitr/RMarkdown/LaTeX.
The end result becomes a pdf suitable for the final course evaluation report.}
\usage{
courseEval(data, syntax = "md")
}
\arguments{
\item{data}{Data frame containing the raw data.}

\item{syntax}{Variable denoting whether LaTeX ("latex") or RMarkdown ("md") syntax is desired. Defaults to "md".}
}
\value{
Table suitable for reformatting and printing with pander. Contains
Ns, means, and SDs for current and prior year, as well as the percentage of
the current year sample responding in each of 7 response categories.
}
\description{
Function for taking in a raw course evaluation dataset from OASIS
and outputting a table of interim values that can be later formatted for use
with knitr/RMarkdown/LaTeX.
The end result becomes a pdf suitable for the final course evaluation report.
}
\examples{
df <- data(sample)
foo <- courseEval(df)
foo
}
\seealso{
\code{\link{welchEMJ}}, \code{\link{pandoc.table}}
}
\keyword{OASIS,}
\keyword{course}
\keyword{evaluation,}
\keyword{knitr}

