% Generated by roxygen2 (4.0.0): do not edit by hand
\name{emphasis}
\alias{emphasis}
\title{Internal function for emphasizing certain words in the row labels for OASIS course evaluation reports.
Used in courseEval.}
\usage{
emphasis(strings, words, syntax = "md")
}
\arguments{
\item{strings}{Array of strings containing words to be emphasized.}

\item{words}{List of words or phrases to emphasize within each string.}

\item{syntax}{Variable denoting whether LaTeX ("latex") or RMarkdown ("md") syntax is desired. Defaults to "md".}
}
\value{
Character vector combining the two values with formatting markup, including a hard line break between.
}
\description{
Internal function for emphasizing certain words in the row labels for OASIS course evaluation reports.
Used in courseEval.
}
\examples{
strings <- c("I hope you have a nice day","I'll see you next week")
words <- list(c("hope","nice"),"you")
emphasis(strings,words,syntax="md")
}
\seealso{
\code{\link{courseEval}}
}
\keyword{LaTeX}
\keyword{OASIS,}
\keyword{RMarkdown,}
\keyword{course}
\keyword{evaluation,}
\keyword{knitr,}

