% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasis_model.R
\docType{data}
\name{example_oasis_df}
\alias{example_oasis_df}
\title{Example \code{data.frame} of MS voxels}
\format{A \code{data.frame} with 100 rows and 13 columns, corresponding to the
predictors of the OASIS model.}
\usage{
example_oasis_df
}
\description{
A \code{data.frame} of MS voxels for OASIS prediction
}
\keyword{datasets}
