% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getComments}
\alias{getComments}
\title{Function for taking in a raw course evaluation dataset from OASIS
and outputting a table of interim values that can be later formatted for use
with knitr/RMarkdown/LaTeX.
The end result becomes a pdf suitable for the final course evaluation report.}
\usage{
getComments(data)
}
\arguments{
\item{data}{Data frame containing the raw data.}
}
\value{
List of data frames containing only the comments.
}
\description{
Function for taking in a raw course evaluation dataset from OASIS
and outputting a table of interim values that can be later formatted for use
with knitr/RMarkdown/LaTeX.
The end result becomes a pdf suitable for the final course evaluation report.
}
\examples{
df <- data(sample)
foo <- getComments(df)
foo
}
\seealso{
\code{\link{courseEval}}
}
\keyword{OASIS,}
\keyword{course}
\keyword{evaluation,}
\keyword{knitr}

