% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oasis_preproc.R
\name{oasis_preproc}
\alias{oasis_preproc}
\title{OASIS Image Preprocessing}
\usage{
oasis_preproc(flair, t1, t2, pd)
}
\arguments{
\item{flair}{flair volume of class nifti}

\item{t1}{flair volume of class nifti}

\item{t2}{flair volume of class nifti}

\item{pd}{flair volume of class nifti}
}
\value{
Returns a list of objects of class nifti: the preprocessed flair, t1, t2, and pd as well as a brain mask.
}
\description{
This function does the required preprocessing for OASIS for the flair, t2,
t1, and pd volumes using fsl through fslr.  The preprcoessing steps are (1) inhomogenity
correct using fsl fast, (2) rigid registration using fsl flirt, (3) creating
a brain mask using fsl BET.  The function returns a list with the inhomogenity corrected
volumes and the brain mask.
}
\examples{
\dontrun{
library(oro.nifti)
flair <- readNIfTI('path/to/flair', reorient = FALSE)
t1 <- readNIfTI('path/to/t1', reorient = FALSE)
t2 <- readNIfTI('path/to/t2', reorient = FALSE)
pd <- readNIfTI('path/to/pd', reorient = FALSE)
oasis_preprocessed_data <- oasis_preproc(flair, t1, t2, pd) }
}

