% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oasis_train_vectors.R
\name{oasis_train_vectors}
\alias{oasis_train_vectors}
\title{OASIS Training Data Frame}
\usage{
oasis_train_vectors(flair, t1, t2, pd, gold_standard, brain_mask = NULL,
  preproc = FALSE, normalize = TRUE, slices = NULL,
  orientation = "axial")
}
\arguments{
\item{flair}{flair volume of class nifti}

\item{t1}{flair volume of class nifti}

\item{t2}{flair volume of class nifti}

\item{pd}{flair volume of class nifti}

\item{brain_mask}{brain mask of class nifti, if NULL a brain mask will be created using fsl BET}

\item{preproc}{calls the oasis_preproc function and performs the necessary preprocessing steps for OASIS}

\item{normalize}{option to perform z-score normalization of the image, should be TRUE unless you are training model
using data alternative normalization}

\item{model}{an object of class glm used to make the OASIS predictions}
}
\description{
This function creates the training vectors from a single study
}

