% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oasis_training.R
\name{oasis_training}
\alias{oasis_training}
\title{OASIS Training}
\usage{
oasis_training(oasis_dataframe)
}
\arguments{
\item{oasis_dataframe}{dataframe from the oasis_train_vectors function}
}
\description{
This function trains the OASIS model from a data frame produced by the function oasis_train_vectors
}

