% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasis_model.R
\docType{data}
\name{original_oasis_model}
\alias{original_oasis_model}
\title{OASIS glm Object}
\format{A \code{glm} object}
\usage{
original_oasis_model
}
\description{
A \code{glm} object containing the OASIS model from the OASIS paper.
}
\keyword{datasets}
