% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_message_caiso.R
\name{system_message_caiso}
\alias{system_message_caiso}
\title{Get CAISO System Messages within a 31 Day Period}
\usage{
system_message_caiso(
  from_date = Sys.Date() - 14,
  to_date = Sys.Date() + 1,
  message_severity = c("ALL", "Emergency", "Urgent", "Normal"),
  resultformat = c("csv", "xml"),
  base_url = "http://oasis.caiso.com/oasisapi",
  tz_hrs_from_gmt = "0700",
  api_version = "1"
)
}
\arguments{
\item{from_date}{A date value defining how far to look back.}

\item{to_date}{A date value defining the end date of the look back period.}

\item{message_severity}{A character vector defining message type (ALL, Emergency, urgent, Normal).}

\item{resultformat}{A character vector defining what type of results to return (csv, xml).}

\item{base_url}{A character vector of Oasis api site.}

\item{tz_hrs_from_gmt}{A 4 digit character vector defining hours from GMT}

\item{api_version}{A character value of API version}
}
\value{
A dataframe of system messages
}
\description{
Get CAISO System Messages within a 31 Day Period
}
\examples{
df <- system_message_caiso()
df <- system_message_caiso(
  from_date = as.Date("2024-03-01"),
  to_date = as.Date("2024-03-07"),
  message_severity = "Emergency"
)
}
