% Generated by roxygen2 (4.0.0): do not edit by hand
\name{twoline}
\alias{twoline}
\title{Internal function for printing two values in the format required for OASIS course evaluation reports.
The first value is unmarked; the second, on a newline, is smaller and italicized.
Used in courseEval.}
\usage{
twoline(var1, var2, syntax)
}
\arguments{
\item{var1}{Variable containing current year data (for unmarked inclusion).}

\item{var2}{Variable containing past year data (for marked inclusion).}

\item{syntax}{Variable denoting whether LaTeX ("latex") or RMarkdown ("md") syntax is desired. Defaults to "md".}
}
\value{
Character vector combining the two values with formatting markup, including a hard line break between.
}
\description{
Internal function for printing two values in the format required for OASIS course evaluation reports.
The first value is unmarked; the second, on a newline, is smaller and italicized.
Used in courseEval.
}
\examples{
N1 <- 100
N2 <- 89
twoline(N1,N2,"latex")
twoline(N1,N2,"md")
}
\seealso{
\code{\link{courseEval}}
}
\keyword{LaTeX}
\keyword{OASIS,}
\keyword{RMarkdown,}
\keyword{course}
\keyword{evaluation,}
\keyword{knitr,}

