% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/voxel_selection.R
\name{voxel_selection}
\alias{voxel_selection}
\title{Voxel Selection Procedure}
\usage{
voxel_selection(flair, brain_mask, cutoff)
}
\arguments{
\item{flair}{flair volume of class nifti}

\item{brain_mask}{a brain mask for the flair}

\item{cutoff}{the percentile cutoff for the thresholding,
passed to \code{\link{quantile}}}
}
\value{
Object of class nifti
}
\description{
This function does the voxel selection
procedure for OASIS.
}

