% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasis_erode.R
\name{oasis_erode}
\alias{oasis_erode}
\title{OASIS Erode Mask}
\usage{
oasis_erode(mask, mm = c(5, 5, 5))
}
\arguments{
\item{mask}{object of class \code{\link{nifti}}}

\item{mm}{Number of erosion (in millimeters)}
}
\value{
Object of class \code{\link{nifti}}
}
\description{
An alternative to using \code{fslerode} for mask erosion of a 
brain mask by a box kernel defined by millimeter
}

