% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oasis_predict.R
\name{oasis_predict}
\alias{oasis_predict}
\title{OASIS Prediction}
\usage{
oasis_predict(flair, t1, t2, pd, brain_mask = NULL, preproc = FALSE,
  normalize = TRUE, model = default_model)
}
\arguments{
\item{flair}{flair volume of class nifti}

\item{t1}{flair volume of class nifti}

\item{t2}{flair volume of class nifti}

\item{pd}{flair volume of class nifti}

\item{brain_mask}{brain mask of class nifti, if NULL a brain mask will be created using fsl BET}

\item{preproc}{calls the oasis_preproc function and performs the necessary preprocessing steps for OASIS}

\item{normalize}{option to perform z-score normalization of the image, should be TRUE unless you train model
using an alternative normalization}

\item{model}{an object of class glm used to make the OASIS predictions}
}
\description{
This function creates the OASIS probability maps.
}

