% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oasis_predict.R
\name{oasis_predict}
\alias{oasis_predict}
\title{OASIS Prediction}
\usage{
oasis_predict(flair, t1, t2, pd, brain_mask = NULL, preproc = FALSE,
  normalize = TRUE, model = NULL)
}
\arguments{
\item{flair}{flair volume of class nifti}

\item{t1}{flair volume of class nifti}

\item{t2}{flair volume of class nifti}

\item{pd}{flair volume of class nifti}

\item{brain_mask}{brain mask of class nifti, if NULL a brain mask will be created using fsl BET}

\item{preproc}{calls the oasis_preproc function and performs the necessary preprocessing steps for OASIS}

\item{normalize}{option to perform z-score normalization of the image, should be TRUE unless you train model
using an alternative normalization}

\item{model}{an object of class glm used to make the OASIS predictions}
}
\value{
Reutrns a volume of class nifti containing the OASIS probability for each voxel.
}
\description{
This function creates the OASIS probability map from a single MRI study with FLAIR, T1, T2, and PD volumes.
}
\examples{
\dontrun{
flair <- readNIfTI('path/to/flair', reorient = FALSE)
t2 <- readNIfTI('path/to/t2', reorient = FALSE)
t1 <- readNIfTI('path/to/t1', reorient = FALSE)
pd <- readNIfTI('path/to/pd', reorient = FALSE)
oasis_map <- oasis_predict(flair = flair, t1 = t1, t2 = t2, pd = pd) }
}

