% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oasis_preproc.R
\name{oasis_preproc}
\alias{oasis_preproc}
\title{OASIS Image Preprocessing}
\usage{
oasis_preproc(flair, t1, t2, pd)
}
\arguments{
\item{flair}{flair volume of class nifti}

\item{t1}{t1 volume of class nifti}

\item{t2}{t2 volume of class nifti}

\item{pd}{pd volume of class nifti}
}
\value{
Returns a list of objects of class nifti, namely the preprocessed FLAIR, T1, T2, and PD registered to the space of the T1 volume.
}
\description{
This function does the required preprocessing for OASIS for the FLAIR, T2,
T1, and PD volumes using fsl through fslr.  The preprcoessing steps are (1) inhomogenity
correct using fsl fast and  (2) rigid registration using fsl flirt to the T1 space.  The function returns
a list with the inhomogenity corrected volumes.
}
\examples{
\dontrun{
library(oro.nifti)
flair <- readNIfTI('path/to/flair', reorient = FALSE)
t1 <- readNIfTI('path/to/t1', reorient = FALSE)
t2 <- readNIfTI('path/to/t2', reorient = FALSE)
pd <- readNIfTI('path/to/pd', reorient = FALSE)
oasis_preprocessed_data <- oasis_preproc(flair, t1, t2, pd) }
}

