% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oasis_training.R
\name{oasis_training}
\alias{oasis_training}
\title{OASIS Training}
\usage{
oasis_training(..., remove_preproc = TRUE)
}
\arguments{
\item{...}{dataframe(s) produced by the oasis_preproc function}

\item{remove_preproc}{a logical stateing if preprocessed volumes need to be removed from the dataframes}
}
\value{
Returns a glm object containing the trained OASIS coefficients to be used by the function oasis_predict.
}
\description{
This function trains the OASIS model from a data frame produced by the function oasis_train_vectors
}
\examples{
\dontrun{
my_oasis_model <- oasis_training(oasis_dataframe_1, oasis_dataframe_2)  }
}

