% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasis_training.R
\name{oasis_training}
\alias{oasis_training}
\title{OASIS Training}
\usage{
oasis_training(..., remove_preproc = FALSE)
}
\arguments{
\item{...}{\code{data.frame}(s) produced by the 
\code{\link{oasis_train_dataframe}} function}

\item{remove_preproc}{a logical stating if preprocessed volumes 
need to be removed from the \code{data.frame}s}
}
\value{
Returns a \code{glm} object containing the trained OASIS 
coefficients to be used by the function \code{\link{oasis_predict}}.
}
\description{
This function trains the OASIS model from a \code{data.frame} 
produced by the function \code{\link{oasis_train_dataframe}}
}
\examples{
\dontrun{
my_oasis_model <- oasis_training(oasis_dataframe_1, oasis_dataframe_2)  
}
}

