% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/voxel_selection.R
\name{voxel_selection}
\alias{voxel_selection}
\title{Voxel Selection Procedure}
\usage{
voxel_selection(flair, brain_mask, cutoff)
}
\arguments{
\item{flair}{flair volume of class nifti}

\item{brain_mask}{brain mask of class nifti}

\item{cutoff}{the percentile cutoff for the thresholding,
passed to \code{\link{quantile}}}
}
\value{
Returns the voxel slection mask as an object of class nifti
}
\description{
This function creates a binary mask for the voxel selection
procedure for OASIS.
}
\examples{
\dontrun{
flair <- readNIfTI('path/to/flair', reorient = FALSE)
brain_mask <- readNIfTI('path/to/brain_mask', reorient = FALSE)
voxel_selection_mask <- oasis_preproc(flair, brain_mask, cutoff = .85) }
}

