% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_selection.R
\name{voxel_selection}
\alias{voxel_selection}
\title{Voxel Selection Procedure}
\usage{
voxel_selection(flair, brain_mask, cutoff)
}
\arguments{
\item{flair}{FLAIR volume of class \code{\link{nifti}}}

\item{brain_mask}{brain mask of class \code{\link{nifti}}}

\item{cutoff}{the percentile cutoff for the thresholding, 
passed to \code{\link{quantile}}}
}
\value{
Returns the voxel selection mask as an object of class \code{\link{nifti}}.
}
\description{
This function creates a binary mask for the voxel selection 
procedure for OASIS.
}
\examples{
\dontrun{
library(fslr)
flair <- readnii('path/to/flair', reorient = FALSE) 
brain_mask <- readnii('path/to/brain_mask', reorient = FALSE) 
voxel_selection_mask <- voxel_selection(flair, brain_mask, cutoff = .85) }
}

